requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.8.1=py37_0
astor=0.8.0=py37_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.0=py37_0
ca-certificates=2019.11.28=hecc5488_0
certifi=2019.11.28=py37_0
chardet=3.0.4=py37_1003
click=7.0=py37_0
cloudpickle=1.2.2=py_0
colorama=0.4.3=py_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=py37_0
cython=0.29.14=py37ha925a31_0
cytoolz=0.10.1=py37he774522_0
dash=1.4.1=py_0
dash-bootstrap-components=0.8.1=py37_0
dash-core-components=1.3.1=py_0
dash-html-components=1.0.1=py_0
dash-renderer=1.1.2=py_0
dash-table=4.4.1=py_0
dask-core=2.9.1=py_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py37_0
flask=1.1.1=py_0
flask-compress=1.4.0=py_0
freetype=2.9.1=ha9979f8_1
future=0.18.2=py37_0
gast=0.2.2=py37_0
google-pasta=0.1.8=py_0
grpcio=1.16.1=py37h351948d_1
h5py=2.9.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
imageio=2.6.1=py37_0
importlib_metadata=1.3.0=py37_0
intel-openmp=2019.4=245
ipykernel=5.1.3=py37h39e3cac_0
ipython=7.11.1=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
itsdangerous=1.1.0=py37_0
jedi=0.15.1=py37_0
jinja2=2.10.3=py_0
joblib=0.14.1=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py37_0
jupyter_client=5.3.4=py37_0
jupyter_core=4.6.1=py37_0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py37_0
keras-gpu=2.2.4=0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37ha925a31_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.11.2=h7bd577a_0
libsodium=1.0.16=h9d3ae62_0
libtiff=4.1.0=h56a325e_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37he774522_0
matplotlib=3.1.1=py37hc8f65d3_0
mistune=0.8.4=py37he774522_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
more-itertools=8.0.2=py_0
msys2-conda-epoch=20160418=1
nbconvert=5.6.1=py37_0
nbformat=4.4.0=py37_0
networkx=2.4=py_0
notebook=6.0.2=py37_0
numpy=1.17.4=py37h4320e6b_0
numpy-base=1.17.4=py37hc3f5095_0
olefile=0.46=py37_0
openssl=1.1.1d=hfa6e2cd_0
opt_einsum=3.1.0=py_0
pandas=0.25.3=py37ha925a31_0
pandas-profiling=1.4.1=py37_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.5.2=py_0
pickleshare=0.7.5=py37_0
pillow=6.2.1=py37hdc69c19_0
pip=19.3.1=py37_0
plotly=4.4.1=py_0
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.2=py_0
protobuf=3.11.2=py37h33f27b4_0
pybase64=1.0.1=pypi_0
pycocotools=2.0=pypi_0
pygments=2.5.2=py_0
pyparsing=2.4.5=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pyrsistent=0.15.6=py37he774522_0
python=3.7.5=h8c8aaf0_0
python-dateutil=2.8.1=py_0
pytz=2019.3=py_0
pywavelets=1.1.1=py37he774522_0
pywin32=227=py37he774522_0
pywinpty=0.5.7=py37_0
pyyaml=5.2=py37he774522_0
pyzmq=18.1.0=py37ha925a31_0
qt=5.9.7=vc14h73c81de_0
retrying=1.3.3=py37_2
scikit-image=0.15.0=py37ha925a31_0
scikit-learn=0.22=py37h6288b17_0
scipy=1.3.2=py37h29ff71c_0
send2trash=1.5.0=py37_0
setuptools=44.0.0=py37_0
sip=4.19.8=py37h6538335_0
six=1.13.0=py37_0
sqlite=3.30.1=he774522_0
tensorboard=2.0.0=pyhb38c66f_1
tensorflow=2.0.0=gpu_py37h57d29ca_0
tensorflow-base=2.0.0=gpu_py37h390e234_0
tensorflow-estimator=2.0.0=pyh2649769_0
tensorflow-gpu=2.0.0=h0d30ee6_0
termcolor=1.1.0=py37_1
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
tk=8.6.8=hfa6e2cd_0
toolz=0.10.0=py_0
tornado=6.0.3=py37he774522_0
traitlets=4.3.3=py37_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
werkzeug=0.16.0=py_0
wheel=0.33.6=py37_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.11.2=py37he774522_0
xz=5.2.4=h2fa13f4_4
yaml=0.1.7=hc54c509_2
zeromq=4.3.1=h33f27b4_3
zipp=0.6.0=py_0
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
